[Common] Move the filesystem.h to common.
