+++ lib/aot/compiler.cpp
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: Apache-2.0
 #include "aot/compiler.h"
+#include "common/filesystem.h"
 #include "common/log.h"
 #include "runtime/instance/memory.h"
 #include "runtime/instance/table.h"
-#include "support/filesystem.h"
 #include <lld/Common/Driver.h>
 #include <llvm/Analysis/TargetLibraryInfo.h>
 #include <llvm/IR/IRBuilder.h>
