+++ lib/loader/filemgr.cpp
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 #include "loader/filemgr.h"
-#include "support/filesystem.h"
+#include "common/filesystem.h"
 
 #include <algorithm>
 #include <iterator>
