+++ tools/ssvm-aot/ssvmc.cpp
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 #include "aot/compiler.h"
+#include "common/filesystem.h"
 #include "loader/loader.h"
 #include "po/argument_parser.h"
-#include "support/filesystem.h"
 #include "validator/validator.h"
 #include <iostream>
