+++ tools/ssvm/main.cpp
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
+#include "common/filesystem.h"
 #include "common/value.h"
 #include "host/wasi/wasimodule.h"
 #include "po/argument_parser.h"
-#include "support/filesystem.h"
 #include "vm/configure.h"
 #include "vm/vm.h"
